python/requirements.txt
openai
langchain
gradio
python-dotenv
